Refactor authentication DTOs and validation logic
Removed `RegisterRequestDto` and `ConfirmEmailRequestDto` from `LoginRequest.cs`.
Created new `ConfirmEmailRequestDto` and `RegisterRequestDto` files with appropriate properties and validation rules.
Updated namespace to `Dentizone.Application.DTOs.Auth` and included necessary using directives.
