+++ Dentizone.Application/DTOs/Auth/ConfirmEmailRequestDto.cs
@@ -1,0 +1,7 @@
+﻿namespace Dentizone.Application.DTOs.Auth;
+
+public class ConfirmEmailRequestDto
+{
+    public string Token { get; set; }
+    public string UserId { get; set; }
+}
